Move PikeState into own module
This change moves the `PikeState` struct definition and implementation
to its own module, and out from the `handler` module. This allows for
easier differentiation between the functionality of `PikeState` and
the `PikeTransactionHandler`.
